add update video service

+++ src/main/java/com/xaqnus/my_tube_backend/video/repository/VideoRepository.java
@@ -7,7 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface VideoRepository extends JpaRepository<Video, Integer> {
+public interface VideoRepository extends JpaRepository<Video, Long> {
 
     List<Video> findAllByUser(User user);
+
+
 }
